+++ src/main/java/com/nemonotfound/NemosCreaturesClient.java
@@ -1,8 +1,9 @@
 package com.nemonotfound;
 
 import com.nemonotfound.client.render.entity.*;
+import com.nemonotfound.client.render.entity.model.ModEntityModelLayers;
 import com.nemonotfound.client.render.entity.model.WildBoarEntityModel;
-import com.nemonotfound.entity.Entities;
+import com.nemonotfound.entity.ModEntityTypes;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -11,9 +12,6 @@ import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
 import net.minecraft.client.model.Dilation;
 import net.minecraft.client.model.TexturedModelData;
 import net.minecraft.client.render.entity.model.*;
-import net.minecraft.util.Identifier;
-
-import static com.nemonotfound.NemosCreatures.MOD_ID;
 
 @Environment(EnvType.CLIENT)
 public class NemosCreaturesClient implements ClientModInitializer {
@@ -23,64 +21,60 @@ public class NemosCreaturesClient implements ClientModInitializer {
     TexturedModelData armorModelData = TexturedModelData.of(ArmorEntityModel.getModelData(ARMOR_DILATION), 64, 32);
     TexturedModelData hatModelData = TexturedModelData.of(ArmorEntityModel.getModelData(HAT_DILATION), 64, 32);
 
-    public static final EntityModelLayer VENOMOUS_SKELETON = new EntityModelLayer(new Identifier(MOD_ID, "venomous_skeleton"), "main");
-    public static final EntityModelLayer VENOMOUS_SKELETON_INNER_ARMOR = new EntityModelLayer(new Identifier(MOD_ID, "venomous_skeleton"), "inner_armor");
-    public static final EntityModelLayer VENOMOUS_SKELETON_OUTER_ARMOR = new EntityModelLayer(new Identifier(MOD_ID, "venomous_skeleton"), "outer_armor");
-    public static final EntityModelLayer VENOMOUS_SKELETON_OUTER = new EntityModelLayer(new Identifier(MOD_ID, "venomous_skeleton"), "outer");
-    public static final EntityModelLayer VENOMOUS_SPIDER = new EntityModelLayer(new Identifier(MOD_ID, "venomous_spider"), "main");
-    public static final EntityModelLayer VENOMOUS_ZOMBIE = new EntityModelLayer(new Identifier(MOD_ID, "venomous_zombie"), "main");
-    public static final EntityModelLayer VENOMOUS_ZOMBIE_INNER_ARMOR = new EntityModelLayer(new Identifier(MOD_ID, "venomous_zombie"), "inner_armor");
-    public static final EntityModelLayer VENOMOUS_ZOMBIE_OUTER_ARMOR = new EntityModelLayer(new Identifier(MOD_ID, "venomous_zombie"), "outer_armor");
-    public static final EntityModelLayer SCORCHED_SKELETON = new EntityModelLayer(new Identifier(MOD_ID, "scorched_skeleton"), "main");
-    public static final EntityModelLayer SCORCHED_SKELETON_INNER_ARMOR = new EntityModelLayer(new Identifier(MOD_ID, "scorched_skeleton"), "inner_armor");
-    public static final EntityModelLayer SCORCHED_SKELETON_OUTER_ARMOR = new EntityModelLayer(new Identifier(MOD_ID, "scorched_skeleton"), "outer_armor");
-    public static final EntityModelLayer SAND_SPIDER = new EntityModelLayer(new Identifier(MOD_ID, "sand_spider"), "main");
-    public static final EntityModelLayer MUMMY = new EntityModelLayer(new Identifier(MOD_ID, "mummy"), "main");
-    public static final EntityModelLayer MUMMY_INNER_ARMOR = new EntityModelLayer(new Identifier(MOD_ID, "mummy"), "inner_armor");
-    public static final EntityModelLayer MUMMY_OUTER_ARMOR = new EntityModelLayer(new Identifier(MOD_ID, "mummy"), "outer_armor");
-    public static final EntityModelLayer CRIMSON_SKELETON = new EntityModelLayer(new Identifier(MOD_ID, "crimson_skeleton"), "main");
-    public static final EntityModelLayer CRIMSON_SKELETON_INNER_ARMOR = new EntityModelLayer(new Identifier(MOD_ID, "crimson_skeleton"), "inner_armor");
-    public static final EntityModelLayer CRIMSON_SKELETON_OUTER_ARMOR = new EntityModelLayer(new Identifier(MOD_ID, "crimson_skeleton"), "outer_armor");
-    public static final EntityModelLayer WARPED_SKELETON = new EntityModelLayer(new Identifier(MOD_ID, "warped_skeleton"), "main");
-    public static final EntityModelLayer WARPED_SKELETON_INNER_ARMOR = new EntityModelLayer(new Identifier(MOD_ID, "warped_skeleton"), "inner_armor");
-    public static final EntityModelLayer WARPED_SKELETON_OUTER_ARMOR = new EntityModelLayer(new Identifier(MOD_ID, "warped_skeleton"), "outer_armor");
-    public static final EntityModelLayer BOAR = new EntityModelLayer(new Identifier(MOD_ID, "wild_boar"), "main");
-
     @Override
     public void onInitializeClient() {
-        EntityRendererRegistry.register(Entities.VENOMOUS_SKELETON, VenomousSkeletonEntityRenderer::new);
-        EntityRendererRegistry.register(Entities.VENOMOUS_SPIDER, VenomousSpiderEntityRenderer::new);
-        EntityRendererRegistry.register(Entities.VENOMOUS_ZOMBIE, VenomousZombieEntityRenderer::new);
-        EntityRendererRegistry.register(Entities.VENOMOUS_CREEPER, VenomousCreeperEntityRenderer::new);
-        EntityRendererRegistry.register(Entities.SCORCHED_SKELETON, ScorchedSkeletonEntityRenderer::new);
-        EntityRendererRegistry.register(Entities.SAND_SPIDER, SandSpiderEntityRenderer::new);
-        EntityRendererRegistry.register(Entities.MUMMY, MummyEntityRenderer::new);
-        EntityRendererRegistry.register(Entities.SCORCHED_CREEPER, ScorchedCreeperEntityRenderer::new);
-        EntityRendererRegistry.register(Entities.CRIMSON_SKELETON, CrimsonSkeletonEntityRenderer::new);
-        EntityRendererRegistry.register(Entities.WARPED_SKELETON, WarpedSkeletonEntityRenderer::new);
-        EntityRendererRegistry.register(Entities.WILD_BOAR, WildBoarEntityRenderer::new);
+        ModEntityModelLayers.registerEntityModelLayers();
 
-        EntityModelLayerRegistry.registerModelLayer(VENOMOUS_SKELETON, SkeletonEntityModel::getTexturedModelData);
-        EntityModelLayerRegistry.registerModelLayer(VENOMOUS_SKELETON_INNER_ARMOR, () -> hatModelData);
-        EntityModelLayerRegistry.registerModelLayer(VENOMOUS_SKELETON_OUTER_ARMOR, () -> armorModelData);
-        EntityModelLayerRegistry.registerModelLayer(VENOMOUS_SKELETON_OUTER, () -> TexturedModelData.of(BipedEntityModel.getModelData(new Dilation(0.25f), 0.0f), 64, 32));
-        EntityModelLayerRegistry.registerModelLayer(VENOMOUS_SPIDER, SpiderEntityModel::getTexturedModelData);
-        EntityModelLayerRegistry.registerModelLayer(VENOMOUS_ZOMBIE, () -> TexturedModelData.of(ZombieEntityModel.getModelData(Dilation.NONE, 0f), 64, 64));
-        EntityModelLayerRegistry.registerModelLayer(VENOMOUS_ZOMBIE_INNER_ARMOR, () -> hatModelData);
-        EntityModelLayerRegistry.registerModelLayer(VENOMOUS_ZOMBIE_OUTER_ARMOR, () -> armorModelData);
-        EntityModelLayerRegistry.registerModelLayer(SCORCHED_SKELETON, SkeletonEntityModel::getTexturedModelData);
-        EntityModelLayerRegistry.registerModelLayer(SCORCHED_SKELETON_INNER_ARMOR, () -> hatModelData);
-        EntityModelLayerRegistry.registerModelLayer(SCORCHED_SKELETON_OUTER_ARMOR, () -> armorModelData);
-        EntityModelLayerRegistry.registerModelLayer(SAND_SPIDER, SpiderEntityModel::getTexturedModelData);
-        EntityModelLayerRegistry.registerModelLayer(MUMMY, () -> TexturedModelData.of(ZombieEntityModel.getModelData(Dilation.NONE, 0f), 64, 64));
-        EntityModelLayerRegistry.registerModelLayer(MUMMY_INNER_ARMOR, () -> hatModelData);
-        EntityModelLayerRegistry.registerModelLayer(MUMMY_OUTER_ARMOR, () -> armorModelData);
-        EntityModelLayerRegistry.registerModelLayer(CRIMSON_SKELETON, SkeletonEntityModel::getTexturedModelData);
-        EntityModelLayerRegistry.registerModelLayer(CRIMSON_SKELETON_INNER_ARMOR, () -> hatModelData);
-        EntityModelLayerRegistry.registerModelLayer(CRIMSON_SKELETON_OUTER_ARMOR, () -> armorModelData);
-        EntityModelLayerRegistry.registerModelLayer(WARPED_SKELETON, SkeletonEntityModel::getTexturedModelData);
-        EntityModelLayerRegistry.registerModelLayer(WARPED_SKELETON_INNER_ARMOR, () -> hatModelData);
-        EntityModelLayerRegistry.registerModelLayer(WARPED_SKELETON_OUTER_ARMOR, () -> armorModelData);
-        EntityModelLayerRegistry.registerModelLayer(BOAR, WildBoarEntityModel::getTexturedModelData);
+        EntityRendererRegistry.register(ModEntityTypes.VENOMOUS_SKELETON, VenomousSkeletonEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.VENOMOUS_SPIDER, VenomousSpiderEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.VENOMOUS_ZOMBIE, VenomousZombieEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.VENOMOUS_CREEPER, VenomousCreeperEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.SCORCHED_SKELETON, ScorchedSkeletonEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.SAND_SPIDER, SandSpiderEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.MUMMY, MummyEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.SCORCHED_CREEPER, ScorchedCreeperEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.CRIMSON_SKELETON, CrimsonSkeletonEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.WARPED_SKELETON, WarpedSkeletonEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.WILD_BOAR, WildBoarEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.SNOWY_SKELETON, SnowySkeletonEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.SNOWY_PIG, SnowyPigEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.SNOWY_COW, SnowyCowEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.SNOWY_SPIDER, SnowySpiderEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.SNOW_SPIDER, SnowSpiderEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.SNOWY_CREEPER, SnowyCreeperEntityRenderer::new);
+        EntityRendererRegistry.register(ModEntityTypes.SNOWY_ZOMBIE, SnowyZombieEntityRenderer::new);
+
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.VENOMOUS_SKELETON, SkeletonEntityModel::getTexturedModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.VENOMOUS_SKELETON_INNER_ARMOR, () -> hatModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.VENOMOUS_SKELETON_OUTER_ARMOR, () -> armorModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.VENOMOUS_SKELETON_OUTER, () -> TexturedModelData.of(BipedEntityModel.getModelData(new Dilation(0.25f), 0.0f), 64, 32));
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.VENOMOUS_SPIDER, SpiderEntityModel::getTexturedModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.VENOMOUS_ZOMBIE, () -> TexturedModelData.of(ZombieEntityModel.getModelData(Dilation.NONE, 0f), 64, 64));
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.VENOMOUS_ZOMBIE_INNER_ARMOR, () -> hatModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.VENOMOUS_ZOMBIE_OUTER_ARMOR, () -> armorModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SCORCHED_SKELETON, SkeletonEntityModel::getTexturedModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SCORCHED_SKELETON_INNER_ARMOR, () -> hatModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SCORCHED_SKELETON_OUTER_ARMOR, () -> armorModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SAND_SPIDER, SpiderEntityModel::getTexturedModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.MUMMY, () -> TexturedModelData.of(ZombieEntityModel.getModelData(Dilation.NONE, 0f), 64, 64));
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.MUMMY_INNER_ARMOR, () -> hatModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.MUMMY_OUTER_ARMOR, () -> armorModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.CRIMSON_SKELETON, SkeletonEntityModel::getTexturedModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.CRIMSON_SKELETON_INNER_ARMOR, () -> hatModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.CRIMSON_SKELETON_OUTER_ARMOR, () -> armorModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.WARPED_SKELETON, SkeletonEntityModel::getTexturedModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.WARPED_SKELETON_INNER_ARMOR, () -> hatModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.WARPED_SKELETON_OUTER_ARMOR, () -> armorModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.BOAR, WildBoarEntityModel::getTexturedModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SNOWY_SKELETON, SkeletonEntityModel::getTexturedModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SNOWY_SKELETON_INNER_ARMOR, () -> hatModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SNOWY_SKELETON_OUTER_ARMOR, () -> armorModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SNOWY_PIG, () -> PigEntityModel.getTexturedModelData(Dilation.NONE));
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SNOWY_COW, CowEntityModel::getTexturedModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SNOWY_SPIDER, SpiderEntityModel::getTexturedModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SNOW_SPIDER, SpiderEntityModel::getTexturedModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SNOWY_ZOMBIE, () -> TexturedModelData.of(ZombieEntityModel.getModelData(Dilation.NONE, 0f), 64, 32));
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SNOWY_ZOMBIE_INNER_ARMOR, () -> hatModelData);
+        EntityModelLayerRegistry.registerModelLayer(ModEntityModelLayers.SNOWY_ZOMBIE_OUTER_ARMOR, () -> armorModelData);
     }
 }