+++ src/main/java/com/nemonotfound/client/render/entity/SandSpiderEntityRenderer.java
@@ -1,5 +1,6 @@
 package com.nemonotfound.client.render.entity;
 
+import com.nemonotfound.client.render.entity.model.ModEntityModelLayers;
 import com.nemonotfound.entity.mob.SandSpiderEntity;
 import net.minecraft.client.render.entity.EntityRendererFactory;
 import net.minecraft.client.render.entity.SpiderEntityRenderer;
@@ -7,7 +8,6 @@ import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.util.Identifier;
 
 import static com.nemonotfound.NemosCreatures.MOD_ID;
-import static com.nemonotfound.NemosCreaturesClient.SAND_SPIDER;
 
 public class SandSpiderEntityRenderer extends SpiderEntityRenderer<SandSpiderEntity> {
 
@@ -15,7 +15,7 @@ public class SandSpiderEntityRenderer extends SpiderEntityRenderer<SandSpiderEnt
     private static final float SCALE = 0.4f;
 
     public SandSpiderEntityRenderer(EntityRendererFactory.Context context) {
-        super(context, SAND_SPIDER);
+        super(context, ModEntityModelLayers.SAND_SPIDER);
         this.shadowRadius *= SCALE;
     }