+++ src/main/java/com/nemonotfound/client/render/entity/WarpedSkeletonEntityRenderer.java
@@ -1,6 +1,7 @@
 package com.nemonotfound.client.render.entity;
 
-import com.nemonotfound.NemosCreaturesClient;
+import com.nemonotfound.client.render.entity.feature.WarpedSkeletonGlowFeatureRenderer;
+import com.nemonotfound.client.render.entity.model.ModEntityModelLayers;
 import net.minecraft.client.render.entity.EntityRendererFactory;
 import net.minecraft.client.render.entity.SkeletonEntityRenderer;
 import net.minecraft.entity.mob.AbstractSkeletonEntity;
@@ -13,8 +14,9 @@ public class WarpedSkeletonEntityRenderer extends SkeletonEntityRenderer {
     private static final Identifier TEXTURE = new Identifier(MOD_ID, "textures/entity/skeleton/warped_skeleton.png");
 
     public WarpedSkeletonEntityRenderer(EntityRendererFactory.Context context) {
-        super(context, NemosCreaturesClient.WARPED_SKELETON,NemosCreaturesClient.WARPED_SKELETON_INNER_ARMOR,
-                NemosCreaturesClient.WARPED_SKELETON_OUTER_ARMOR);
+        super(context, ModEntityModelLayers.WARPED_SKELETON, ModEntityModelLayers.WARPED_SKELETON_INNER_ARMOR,
+                ModEntityModelLayers.WARPED_SKELETON_OUTER_ARMOR);
+        this.addFeature(new WarpedSkeletonGlowFeatureRenderer<>(this));
     }
 
     @Override