+++ changelog.md
@@ -1,6 +1,21 @@
-# Changelog v1.6.4
+# Changelog v1.7
+
+## Additions
+- Added Snowy mobs
+  - Snowy Skeleton
+  - Snowy Spider
+  - Snow Spider
+  - Snowy Creeper
+  - Snowy Zombie
+  - Snowy Pig
+  - Snowy Cow
 
 ## Changes
-- Vanilla mobs will not be replaced in specific biomes. Instead, we now have a variety of mobs
+- Crimson & Warped Skeleton glow now
+- Creeper effect clouds give a level one effect now
+- Creeper effects last 10 seconds
+
+## Fixes
+- Venomous Skeleton's "pants" are incomplete between legs
 
 For more information check the [wiki](https://www.nemonotfound.com/minecraft-mods/nemos-creatures).